requirements.txt
# Core dependencies
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-google-genai
langchain-anthropic
faiss-cpu

# For Streamlit UI (app.py)
streamlit

# For FastAPI (api.py)
fastapi
uvicorn[standard]
pydantic

# Other dependencies
python-dotenv
